chore: add tests and refactor
