+++ airwallex/src/main/java/com/airwallex/android/view/inputs/ValidatedInput.kt
@@ -1,0 +1,7 @@
+package com.airwallex.android.view.inputs
+
+internal interface ValidatedInput {
+    val isValid: Boolean
+    val emptyErrorMessage: String
+    val invalidErrorMessage: String
+}
