+++ airwallex/src/main/java/com/airwallex/android/view/util/CountryUtils.kt
@@ -1,5 +1,6 @@
-package com.airwallex.android.view
+package com.airwallex.android.view.util
 
+import com.airwallex.android.view.CountryAutoCompleteView
 import java.util.*
 
 internal object CountryUtils {
