+++ airwallex/src/main/java/com/airwallex/android/view/util/ExpiryDateUtils.kt
@@ -1,4 +1,4 @@
-package com.airwallex.android.view
+package com.airwallex.android.view.util
 
 import androidx.annotation.Size
 import java.util.*
