+++ airwallex/src/main/res/layout/widget_card.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content">
 
-    <com.airwallex.android.view.CardNumberTextInputLayout
+    <com.airwallex.android.view.inputs.CardNumberTextInputLayout
         android:id="@+id/atlCardNumber"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -13,7 +13,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <com.airwallex.android.view.CardNameTextInputLayout
+    <com.airwallex.android.view.inputs.CardNameTextInputLayout
         android:id="@+id/atlCardName"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -23,7 +23,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/atlCardNumber" />
 
-    <com.airwallex.android.view.CardExpiryTextInputLayout
+    <com.airwallex.android.view.inputs.CardExpiryTextInputLayout
         android:id="@+id/atlCardExpiry"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
@@ -34,7 +34,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/atlCardName" />
 
-    <com.airwallex.android.view.CardCvcTextInputLayout
+    <com.airwallex.android.view.inputs.CardCvcTextInputLayout
         android:id="@+id/atlCardCvc"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
@@ -45,7 +45,7 @@
         app:layout_constraintStart_toEndOf="@id/atlCardExpiry"
         app:layout_constraintTop_toBottomOf="@id/atlCardName" />
 
-    <com.airwallex.android.view.EmailTextInputLayout
+    <com.airwallex.android.view.inputs.EmailTextInputLayout
         android:id="@+id/atlCardEmail"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
