+++ airwallex/src/test/java/com/airwallex/android/view/CardCvcEditTextTest.kt
@@ -3,6 +3,7 @@ package com.airwallex.android.view
 import androidx.appcompat.view.ContextThemeWrapper
 import androidx.test.core.app.ApplicationProvider
 import com.airwallex.android.ui.R
+import com.airwallex.android.view.inputs.CardCvcEditText
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 import kotlin.test.Test
