+++ airwallex/src/test/java/com/airwallex/android/view/CardExpiryTextInputLayoutTest.kt
@@ -3,6 +3,8 @@ package com.airwallex.android.view
 import androidx.appcompat.view.ContextThemeWrapper
 import androidx.test.core.app.ApplicationProvider
 import com.airwallex.android.R
+import com.airwallex.android.view.inputs.CardExpiryEditText
+import com.airwallex.android.view.inputs.CardExpiryTextInputLayout
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
