+++ airwallex/src/test/java/com/airwallex/android/view/CardUtilsTest.kt
@@ -1,5 +1,6 @@
 package com.airwallex.android.view
 
+import com.airwallex.android.view.util.CardUtils
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
