+++ airwallex/src/test/java/com/airwallex/android/view/EmailTextInputLayoutTest.kt
@@ -3,7 +3,7 @@ package com.airwallex.android.view
 import androidx.appcompat.view.ContextThemeWrapper
 import androidx.test.core.app.ApplicationProvider
 import com.airwallex.android.R
-import com.airwallex.android.view.inputs.CardNameTextInputLayout
+import com.airwallex.android.view.inputs.EmailTextInputLayout
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
@@ -12,33 +12,34 @@ import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
 @RunWith(RobolectricTestRunner::class)
-class CardNameTextInputLayoutTest {
+class EmailTextInputLayoutTest {
+
     private val context = ContextThemeWrapper(
         ApplicationProvider.getApplicationContext(),
         R.style.AirwallexDefaultTheme
     )
 
-    private val inputLayout: CardNameTextInputLayout by lazy {
-        CardNameTextInputLayout(context, null)
+    private val inputLayout: EmailTextInputLayout by lazy {
+        EmailTextInputLayout(context, null)
     }
 
     @Test
     fun `test error messages`() {
         assertEquals(
             inputLayout.emptyErrorMessage,
-            context.getString(R.string.airwallex_empty_card_name)
+            context.getString(R.string.airwallex_empty_email)
         )
         assertEquals(
             inputLayout.invalidErrorMessage,
-            context.getString(R.string.airwallex_empty_card_name)
+            context.getString(R.string.airwallex_invalid_email)
         )
     }
 
     @Test
     fun `test isValid`() {
-        inputLayout.teInput.setText("Name")
+        inputLayout.teInput.setText("john.doe@airwallex.com")
         assertTrue(inputLayout.isValid)
-        inputLayout.teInput.setText("")
+        inputLayout.teInput.setText("abc@de")
         assertFalse(inputLayout.isValid)
     }
 }
